refactor: restructure package initializer
