+++ src/__init__.py
@@ -1,0 +1,19 @@
+"""lawdigest package initialization."""
+
+from .data_operations_OOP import (
+    DatabaseManager,
+    DataFetcher,
+    DataProcessor,
+    AISummarizer,
+    APISender,
+    WorkFlowManager,
+)
+
+__all__ = [
+    "DatabaseManager",
+    "DataFetcher",
+    "DataProcessor",
+    "AISummarizer",
+    "APISender",
+    "WorkFlowManager",
+]
